backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
requests
sentence-transformers
faiss-cpu
python-dotenv